added error handling to controller

+++ App/controllers/user.py
@@ -6,11 +6,12 @@ from sqlalchemy.exc import IntegrityError
 
 
 def create_user(userId, firstname, lastname, username, email, password):
+    newUser = User(userId=userId, firstname=firstname, lastname=lastname,
+                   username=username, email=email, password=password)
     try:
-        newuser = User(userId=userId, firstname=firstname, lastname=lastname,
-                       username=username, email=email, password=password)
-        db.session.add(newuser)
-        db.session.commit()
+        if newUser:
+            db.session.add(newUser)
+            db.session.commit()
     except IntegrityError:
         db.session.rollback()
         return 'ERROR: Failed to create new staff'
@@ -23,19 +24,14 @@ def get_user(id):
 
 
 def get_all_users():
-    try:
-        return User.query.all()
-    except:
-        return 'ERROR: Failed to get all staffs'
+    return User.query.all()
 
 
 def get_all_users_toDict():
     users = get_all_users()
     try:
-        if not users:
-            return []
-        users = [user.toDict() for user in users]
-        return users
+        if users:
+            return [user.toDict() for user in users]
     except:
         return 'ERROR: Failed to get all staffs in dictionary format'
 
@@ -60,10 +56,12 @@ def updateUser(userId, username, firstname, lastname, email, password):
 
 
 def deleteUser(userId):
+    user = get_user(userId)
     try:
-        user = get_user(userId)
-        db.session.delete(user)
-        db.session.commit()
+        if user:
+            user = get_user(userId)
+            db.session.delete(user)
+            db.session.commit()
     except:
         db.session.rollback()
         return 'ERROR: Failed to delete the student'
